requirements.txt
# Core dependencies
torch>=2.1.0
transformers>=4.36.0
accelerate>=0.25.0
datasets>=2.14.0
einops>=0.7.0
wandb>=0.16.0

# Diffusion specific
diffusers>=0.24.0
tqdm>=4.66.0
scipy>=1.11.0

# Optimization
bitsandbytes>=0.41.0  # For 8-bit Adam
peft>=0.7.0  # For LoRA/QLoRA
deepspeed>=0.12.0  # For multi-GPU training

# Evaluation
nltk>=3.8.0
rouge-score>=0.1.2
sacrebleu>=2.3.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.1.0

# Configuration
PyYAML>=6.0.0